i am a bitch.
i'm not sure if this is ok or not but i found
someones example online and parsed the code. i
found what i was doing wrong with the squares
rezing thing and i was both on the right and wrong
track. i will try to parse the rest of their code
to fix my creating a new board problem but i will
try to spend more time actually studying what i
was doing wrong and learn from this.